thread_sync/primes.c: Prefer ULL over size_t
size_t is needed only when addressing memory (e.g. when accessing
array members). This is not the case here. Thus prefer unsigned
long long so that we can check for ULONG_MAX.

+++ thread_sync/primes.c
@@ -17,14 +17,14 @@
 unsigned long
 prime_get_nth(unsigned long n)
 {
-    size_t i;
+    unsigned long long i;
 
     /* The dumbest and possibly grossest way to calculate primes with even
      * dumber overflow protection. */
     n++;
 
     for (i = 2; n && i >= 2; i++) {
-        size_t j;
+        unsigned long long j;
 
         for (j = 2; j < i; j++) {
             if (i % j == 0)
